refactor: keep runtime staging in one tested module
