+++ packages/extension/scripts/runtimeStaging/model.ts
@@ -1,13 +1,108 @@
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 import { createRequire } from 'node:module';
-import {
-  RUNTIME_TARGET_CONFIG,
-  type ExtensionRuntimeTarget,
-} from './types';
 
 const require = createRequire(import.meta.url);
 const requireFromCore = createRequire(require.resolve('@codegraphy-dev/core'));
+
+export const EXTENSION_RUNTIME_TARGETS = [
+  'linux-x64',
+  'darwin-arm64',
+  'win32-x64',
+] as const;
+
+export type ExtensionRuntimeTarget = typeof EXTENSION_RUNTIME_TARGETS[number];
+
+type RuntimeTargetConfig = {
+  platform: NodeJS.Platform;
+  arch: string;
+  libsqlPackageName: string;
+  esbuildPackageName: string;
+  parcelWatcherPackageName: string;
+  nativePrebuildDirectory: string;
+};
+
+export const RUNTIME_TARGET_CONFIG = {
+  'linux-x64': {
+    platform: 'linux',
+    arch: 'x64',
+    libsqlPackageName: '@libsql/linux-x64-gnu',
+    esbuildPackageName: '@esbuild/linux-x64',
+    parcelWatcherPackageName: '@parcel/watcher-linux-x64-glibc',
+    nativePrebuildDirectory: 'linux-x64',
+  },
+  'darwin-arm64': {
+    platform: 'darwin',
+    arch: 'arm64',
+    libsqlPackageName: '@libsql/darwin-arm64',
+    esbuildPackageName: '@esbuild/darwin-arm64',
+    parcelWatcherPackageName: '@parcel/watcher-darwin-arm64',
+    nativePrebuildDirectory: 'darwin-arm64',
+  },
+  'win32-x64': {
+    platform: 'win32',
+    arch: 'x64',
+    libsqlPackageName: '@libsql/win32-x64-msvc',
+    esbuildPackageName: '@esbuild/win32-x64',
+    parcelWatcherPackageName: '@parcel/watcher-win32-x64',
+    nativePrebuildDirectory: 'win32-x64',
+  },
+} satisfies Record<ExtensionRuntimeTarget, RuntimeTargetConfig>;
+
+type RuntimeTargetOptions = {
+  environment?: Partial<Pick<NodeJS.ProcessEnv, 'CODEGRAPHY_VSIX_TARGETS'>>;
+  platform?: NodeJS.Platform;
+  arch?: string;
+};
+
+function isExtensionRuntimeTarget(value: string): value is ExtensionRuntimeTarget {
+  return EXTENSION_RUNTIME_TARGETS.includes(value as ExtensionRuntimeTarget);
+}
+
+export function resolveExtensionRuntimeTarget({
+  environment = process.env,
+  platform = process.platform,
+  arch = process.arch,
+}: RuntimeTargetOptions = {}): ExtensionRuntimeTarget {
+  const requestedTargets = environment.CODEGRAPHY_VSIX_TARGETS
+    ?.split(',')
+    .map(target => target.trim())
+    .filter(Boolean);
+
+  if (requestedTargets && requestedTargets.length !== 1) {
+    throw new Error('Extension runtime staging requires exactly one VSIX target.');
+  }
+
+  const hostTarget = EXTENSION_RUNTIME_TARGETS.find((target) => {
+    const config = RUNTIME_TARGET_CONFIG[target];
+    return config.platform === platform && config.arch === arch;
+  });
+  if (!hostTarget) {
+    throw new Error(
+      `Unsupported extension runtime host: ${platform}-${arch}. `
+      + `Supported targets are ${EXTENSION_RUNTIME_TARGETS.join(', ')}.`,
+    );
+  }
+
+  const requestedTarget = requestedTargets?.[0];
+  if (!requestedTarget) {
+    return hostTarget;
+  }
+  if (!isExtensionRuntimeTarget(requestedTarget)) {
+    throw new Error(
+      `Unsupported extension runtime target: ${requestedTarget}. `
+      + `Supported targets are ${EXTENSION_RUNTIME_TARGETS.join(', ')}.`,
+    );
+  }
+  if (requestedTarget !== hostTarget) {
+    throw new Error(
+      `Cannot stage ${requestedTarget} native files on ${hostTarget}. `
+      + 'Build the VSIX on its matching host.',
+    );
+  }
+
+  return requestedTarget;
+}
 
 const TREE_SITTER_GRAMMAR_PACKAGE_NAMES = [
   'tree-sitter-c',
@@ -224,4 +319,81 @@ export function createRuntimePackagePlans(
       config.nativePrebuildDirectory,
     )),
   ];
+}
+
+export function getVendoredPackageRootPath(
+  outputFilePath: string,
+  packageName: string,
+): string {
+  return path.join(path.dirname(outputFilePath), 'node_modules', ...packageName.split('/'));
+}
+
+function toPackageRelativeEntrypoint(entrypoint: string): string {
+  return entrypoint.endsWith('/')
+    ? `${entrypoint}index.js`
+    : `${entrypoint}/index.js`;
+}
+
+function normalizeVendoredPackageEntrypoint(packageRootPath: string): void {
+  const manifestPath = path.join(packageRootPath, 'package.json');
+  const manifest = JSON.parse(fs.readFileSync(manifestPath, 'utf8')) as {
+    main?: unknown;
+    [key: string]: unknown;
+  };
+  if (typeof manifest.main !== 'string' || path.extname(manifest.main) !== '') {
+    return;
+  }
+
+  const normalizedMain = toPackageRelativeEntrypoint(manifest.main);
+  if (!fs.existsSync(path.join(packageRootPath, normalizedMain))) {
+    return;
+  }
+
+  fs.writeFileSync(
+    manifestPath,
+    `${JSON.stringify({ ...manifest, main: normalizedMain }, null, 2)}\n`,
+  );
+}
+
+export function copyRuntimePackage(
+  outputFilePath: string,
+  packageName: string,
+  relativeFilePaths: readonly string[],
+  resolvePackageRootPath: (packageName: string) => string = resolveRuntimePackageRootPath,
+): string {
+  const sourcePath = resolvePackageRootPath(packageName);
+  const targetPath = getVendoredPackageRootPath(outputFilePath, packageName);
+
+  fs.rmSync(targetPath, { recursive: true, force: true });
+
+  for (const relativeFilePath of [...new Set(relativeFilePaths)]) {
+    const sourceFilePath = path.join(sourcePath, relativeFilePath);
+    if (!fs.existsSync(sourceFilePath) || !fs.statSync(sourceFilePath).isFile()) {
+      throw new Error(`Runtime package file is not a file: ${packageName}/${relativeFilePath}`);
+    }
+
+    const targetFilePath = path.join(targetPath, relativeFilePath);
+    fs.mkdirSync(path.dirname(targetFilePath), { recursive: true });
+    fs.cpSync(sourceFilePath, targetFilePath, { force: true });
+  }
+
+  normalizeVendoredPackageEntrypoint(targetPath);
+  return targetPath;
+}
+
+export function syncExtensionRuntimePackages(
+  outputFilePath: string,
+  target: ExtensionRuntimeTarget = resolveExtensionRuntimeTarget(),
+): string[] {
+  fs.rmSync(path.join(path.dirname(outputFilePath), 'node_modules'), {
+    recursive: true,
+    force: true,
+  });
+
+  return createRuntimePackagePlans(target).map(plan => copyRuntimePackage(
+    outputFilePath,
+    plan.packageName,
+    plan.relativeFilePaths,
+    plan.resolvePackageRootPath,
+  ));
 }
